chore: rewrite library to a full implementation
