--- component.mk
@@ -1,2 +1,0 @@
-COMPONENT_ADD_INCLUDEDIRS := ./src src/platforms/esp/32
-COMPONENT_SRCDIRS := ./src src/platforms/esp/32
