+++ platformio.ini
@@ -13,3 +13,6 @@ platform = espressif8266
 board = esp01
 framework = arduino
 lib_deps =
+    bblanchon/ArduinoJson@^6.20.1
+    ottowinter/ESPAsyncTCP-esphome
+    ottowinter/ESPAsyncWebServer-esphome
