+++ src/AlertEffect.h
@@ -1,0 +1,8 @@
+#pragma once
+
+enum AlertEffect
+{
+    ALERT_EFFECT_NONE,
+    ALERT_EFFECT_SELECT,
+    ALERT_EFFECT_LSELECT,
+};
