+++ src/ColorMode.h
@@ -1,0 +1,8 @@
+#pragma once
+
+enum ColorMode
+{
+    COLOR_MODE_HS,
+    COLOR_MODE_XY,
+    COLOR_MODE_CT
+};
