+++ src/LightEffect.h
@@ -1,0 +1,7 @@
+#pragma once
+
+enum LightEffect
+{
+    LIGHT_EFFECT_NONE,
+    LIGHT_EFFECT_COLORLOOP,
+};
