+++ src/XYPoint.h
@@ -1,0 +1,7 @@
+#pragma once
+
+struct XYPoint
+{
+    float x;
+    float y;
+};
